+++ src/lang/ADempiere/es/form/index.js
@@ -16,16 +16,17 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-import accountingViewer from './accountingViewer'
+import VBankStatementMatch from './VBankStatementMatch'
 import mnemonicCommand from './mnemonicCommand.js'
 import expressShipment from './expressShipment.js'
 import expressMovement from './expressMovement.js'
 import workflowActivity from './workflowActivity'
+import accountingViewer from './accountingViewer'
 import expressReceipt from './expressReceipt.js'
 import priceChecking from './priceChecking'
+import pointOfSales from './pointOfSales'
 import productInfo from './productInfo'
 import VAllocation from './VAllocation'
-import VBankStatementMatch from './VBankStatementMatch'
 import VFileImport from './VFileImport'
 import timeRecord from './timeRecord'
 import match from './match'
@@ -42,6 +43,7 @@ export default {
   productInfo,
   VAllocation,
   VBankStatementMatch,
+  pointOfSales,
   VFileImport,
   timeRecord,
   match,
